refactor: split StatusBarManager + skip redundant menu-bar renders (T12/T15)
T12: StatusBarManager.swift (731 lines) split into focused extension files
per the v2.4.0 UsageViewModel precedent — pure moves, no behavior change:
- StatusBarManager.swift: stored state, setup, deinit (core lifecycle)
- +ButtonUpdate: updateButton + star color/sparkle/render state
- +Countdown: adaptive ticker + pure countdownResetDate(for:now:)
- +Panel: toggle/positioning + PopoverPanel/PopoverContentView/
  TransparentHostingView (private -> internal for cross-file visibility;
  shared stored state de-privatized like UsageViewModel's split)
Micro-DRY bundled per the review: new AppPaths.applicationSupport()
replaces the fatalError guard duplicated in SingleInstanceGuard +
TokenLedger; buttonScreenRect(for:) collapses the duplicated button-rect
conversion in positionPanel/currentTopAnchor. Dropped the dead instance
countdownResetDate(for:) wrapper (no callers).

T15: updateButton now keys the render on MenuBarRenderKey (text,
whole-percent bucket, color, isBroken, isSparkle, appearance name) and
skips the NSAttributedString+NSImage rebuild when unchanged — countdown
ticks fire every 10s but the compact text changes ~once a minute, so ~85%
of ticks were redundant allocations. Countdown timer bookkeeping always
runs (it manages timers, not pixels). (Decided D15: key-skip, NOT a
combined-image cache.)

4 new MenuBarRenderKey tests (1038 -> 1042 in 71 suites);
spec/ARCHITECTURE.md tree + spec/DATA_LAYER.md, README coverage,
CHANGELOG updated.
