+++ AIBattery/Services/SingleInstanceGuard.swift
@@ -10,12 +10,8 @@ import AppKit
 /// bundle identifier, which cleans up zombie processes that may have survived
 /// a crash or force-quit (preventing RBSRequestErrorDomain Code=5).
 public enum SingleInstanceGuard {
-    private static let lockPath: String = {
-        guard let appSupport = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first else {
-            fatalError("Application Support directory unavailable")
-        }
-        return appSupport.appendingPathComponent("AIBattery/aibattery.lock").path
-    }()
+    private static let lockPath: String = AppPaths.applicationSupport()
+        .appendingPathComponent("aibattery.lock").path
 
     /// File descriptor for the lock file — kept open for the process lifetime.
     /// Process-wide state set at startup before any concurrency is active and treated
