+++ AIBattery/Services/TokenLedger.swift
@@ -154,12 +154,7 @@ final class TokenLedger: @unchecked Sendable {
     // MARK: - Storage
 
     private static var defaultFileURL: URL {
-        guard let appSupport = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first else {
-            fatalError("Application Support directory unavailable")
-        }
-        let dir = appSupport.appendingPathComponent("AIBattery")
-        try? FileManager.default.createDirectory(at: dir, withIntermediateDirectories: true)
-        return dir.appendingPathComponent("token-ledger.json")
+        AppPaths.applicationSupport().appendingPathComponent("token-ledger.json")
     }
 
     private static func load(from url: URL) -> LedgerData {
